+++ baybe/parameters/base.py
@@ -13,11 +13,8 @@ from baybe.parameters.enum import ParameterEncoding
 from baybe.utils import SerialMixin, get_base_structure_hook, unstructure_base
 from baybe.utils.serialization import converter
 
-# TODO: Reactive slots in all classes once cached_property is supported:
-#   https://github.com/python-attrs/attrs/issues/164
 
-
-@define(frozen=True, slots=False)
+@define(frozen=True)
 class Parameter(ABC, SerialMixin):
     """Abstract base class for all parameters.
 
@@ -48,11 +45,9 @@ class Parameter(ABC, SerialMixin):
         """
 
 
-@define(frozen=True, slots=False)
+@define(frozen=True)
 class DiscreteParameter(Parameter, ABC):
     """Abstract class for discrete parameters."""
-
-    # TODO [15280]: needs to be refactored
 
     # class variables
     is_discrete: ClassVar[bool] = True