+++ baybe/parameters/categorical.py
@@ -13,7 +13,7 @@ from baybe.parameters.enum import CategoricalEncoding
 from baybe.parameters.validation import validate_unique_values
 
 
-@define(frozen=True, slots=False)
+@define(frozen=True)
 class CategoricalParameter(DiscreteParameter):
     """Parameter class for categorical parameters."""
 
@@ -55,7 +55,7 @@ class CategoricalParameter(DiscreteParameter):
         return comp_df
 
 
-@define(frozen=True, slots=False)
+@define(frozen=True)
 class TaskParameter(CategoricalParameter):
     """Parameter class for task parameters."""